// web/src/api/generated/schemas/reviewCardResource.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fonzytooter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod'
import { RatingPreviewResource } from './ratingPreviewResource.zod'

export const ReviewCardResource = zod.strictObject({
  $schema: zod.url().optional().describe('A URL to the JSON Schema for this object.'),
  answer: zod.string(),
  courseId: zod.string(),
  due: zod.boolean(),
  dueAt: zod.iso.datetime({ offset: true }),
  hint: zod.string(),
  id: zod.string(),
  lastReviewedAt: zod.iso.datetime({ offset: true }).optional(),
  moduleId: zod.string(),
  objectiveIds: zod.array(zod.string()),
  order: zod.int(),
  previews: zod.array(RatingPreviewResource),
  prompt: zod.string(),
  sourceLessonId: zod.string(),
  state: zod.enum(['new', 'learning', 'review', 'relearning']),
  virtual: zod.boolean(),
})

export type ReviewCardResource = zod.input<typeof ReviewCardResource>
export type ReviewCardResourceOutput = zod.output<typeof ReviewCardResource>